chore: update quizzes data
